style: run ktfmt format
