+++ app/src/main/java/com/github/onlynotesswent/ui/common/FolderComposables.kt
@@ -6,17 +6,12 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.draganddrop.dragAndDropSource
 import androidx.compose.foundation.draganddrop.dragAndDropTarget
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -34,7 +29,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
 import com.github.onlynotesswent.R
 import com.github.onlynotesswent.model.common.Visibility
 import com.github.onlynotesswent.model.folder.Folder
@@ -165,55 +159,8 @@ fun FolderDialog(
     onDismiss: () -> Unit,
     onConfirm: (String, Visibility) -> Unit,
     action: String,
-    oldVis: Visibility? = null,
+    oldVisibility: Visibility? = null,
     oldName: String = ""
 ) {
-
-  var name by remember { mutableStateOf(oldName) }
-  var visibility: Visibility? by remember { mutableStateOf(oldVis) }
-  var expandedVisibility by remember { mutableStateOf(false) }
-
-  Dialog(onDismissRequest = onDismiss) {
-    Card(colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)) {
-      Column(
-          modifier = Modifier.padding(16.dp).testTag("folderDialog"),
-          verticalArrangement = Arrangement.spacedBy(8.dp),
-          horizontalAlignment = Alignment.CenterHorizontally) {
-            Row(
-                modifier = Modifier.fillMaxWidth(0.92f),
-                horizontalArrangement = Arrangement.Start) {
-                  Text("$action Folder", style = MaterialTheme.typography.titleLarge)
-                }
-
-            OutlinedTextField(
-                value = name,
-                onValueChange = { name = Folder.formatName(it) },
-                label = { Text("Folder Name") },
-                modifier = Modifier.testTag("inputFolderName"))
-
-            OptionDropDownMenu(
-                value = visibility?.toReadableString() ?: "Choose visibility",
-                expanded = expandedVisibility,
-                buttonTag = "visibilityButton",
-                menuTag = "visibilityMenu",
-                onExpandedChange = { expandedVisibility = it },
-                items = Visibility.READABLE_STRINGS,
-                onItemClick = { visibility = Visibility.fromReadableString(it) },
-                modifier = Modifier.testTag("visibilityDropDown"),
-                widthFactor = 0.94f)
-
-            Row(modifier = Modifier.fillMaxWidth(0.92f), horizontalArrangement = Arrangement.End) {
-              Button(onClick = onDismiss, modifier = Modifier.testTag("dismissFolderAction")) {
-                Text("Cancel")
-              }
-              Button(
-                  enabled = name.isNotEmpty() && visibility != null,
-                  onClick = { onConfirm(name, visibility ?: Visibility.DEFAULT) },
-                  modifier = Modifier.testTag("confirmFolderAction")) {
-                    Text(action)
-                  }
-            }
-          }
-    }
-  }
+  CreationDialog(onDismiss, onConfirm, action, oldVisibility, oldName, "Folder")
 }
