+++ app/src/main/java/com/github/onlynotesswent/ui/common/NoteComposables.kt
@@ -40,7 +40,6 @@ import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.github.onlynotesswent.model.common.Visibility
-import com.github.onlynotesswent.model.folder.Folder
 import com.github.onlynotesswent.model.folder.FolderViewModel
 import com.github.onlynotesswent.model.note.Note
 import com.github.onlynotesswent.model.note.NoteViewModel
@@ -195,52 +194,9 @@ fun NoteDialog(
     oldVisibility: Visibility? = Visibility.PRIVATE,
     oldName: String = ""
 ) {
-  var name by remember { mutableStateOf(oldName) }
-  var visibility: Visibility? by remember { mutableStateOf(oldVisibility) }
-  var expandedVisibility by remember { mutableStateOf(false) }
-
-  androidx.compose.ui.window.Dialog(onDismissRequest = onDismiss) {
-    Card(colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)) {
-      Column(
-          modifier = Modifier.padding(16.dp).testTag("noteDialog"),
-          verticalArrangement = Arrangement.spacedBy(8.dp),
-          horizontalAlignment = Alignment.CenterHorizontally) {
-            Row(
-                modifier = Modifier.fillMaxWidth(0.92f),
-                horizontalArrangement = Arrangement.Start) {
-                  Text("$action Note", style = MaterialTheme.typography.titleLarge)
-                }
-            OutlinedTextField(
-                value = name,
-                onValueChange = { name = Folder.formatName(it) },
-                label = { Text("Note Name") },
-                modifier = Modifier.testTag("inputNoteName"))
-            OptionDropDownMenu(
-                value = visibility?.toReadableString() ?: "Choose visibility",
-                expanded = expandedVisibility,
-                buttonTag = "visibilityButton",
-                menuTag = "visibilityMenu",
-                onExpandedChange = { expandedVisibility = it },
-                items = Visibility.READABLE_STRINGS,
-                onItemClick = { visibility = Visibility.fromReadableString(it) },
-                modifier = Modifier.testTag("visibilityDropDown"),
-                widthFactor = 0.94f)
-
-            Row(modifier = Modifier.fillMaxWidth(0.92f), horizontalArrangement = Arrangement.End) {
-              Button(onClick = onDismiss, modifier = Modifier.testTag("dismissNoteAction")) {
-                Text("Cancel")
-              }
-              Button(
-                  enabled = name.isNotEmpty() && visibility != null,
-                  onClick = { onConfirm(name, visibility ?: Visibility.DEFAULT) },
-                  modifier = Modifier.testTag("confirmNoteAction")) {
-                    Text(action)
-                  }
-            }
-          }
-    }
-  }
+  CreationDialog(onDismiss, onConfirm, action, oldVisibility, oldName, "Note")
 }
+
 /**
  * A composable function that displays an `OutlinedTextField` with an optional trailing icon.
  *
